chore(bot): split up implementation
* split up concerns in multiple modules
* allow traces to be easily recorded; use the
  RECORD_TRACE=name_of_trace env variable to
  configure recording upon start

Fixed bugs:

* ensure conclusion=neutral checks are being
  not treated as successful
* early reject check_suite event processing,
  unless conclusion=success
* compute effective review when evaluating
  human approval
* correct error handling
